move workflows database accessors to their own file
